final version

+++ src/main/java/org/example/restassignment/services/StudentDeleteOrUpdateOrFindService.java
@@ -1,5 +1,6 @@
 package org.example.restassignment.services;
 
+import org.example.restassignment.errors.StudentNotFound;
 import org.example.restassignment.models.DTOs.RequestDTOs.RequestUpdate;
 import org.example.restassignment.models.DTOs.ResponseDTOs.FoundStudent;
 import org.example.restassignment.models.Entities.Student;
@@ -11,7 +12,8 @@ public class StudentDeleteOrUpdateOrFindService {
     public FoundStudent  findStudnet(int id) {
      Student student=   studentRepository.getByID(id);
      if(student==null){
-         return null;
+               throw new StudentNotFound("Student not found");
+
      }
      FoundStudent foundStudent=new FoundStudent(
              student.getId(),student.getName(),student.getAge(),student.getScore()
@@ -21,12 +23,20 @@ public class StudentDeleteOrUpdateOrFindService {
 
     public void  deleteStudent(int id) {
 
+        Student ToBeUpdateleted = studentRepository.getByID(id);
+
+        if( ToBeUpdateleted==null){
+            throw new StudentNotFound("Student not found");
+        }
             studentRepository.delete(id);
 
     }
     public  void  updateStudent(@NonNull RequestUpdate  requestUpdate) {
 
         Student ToBeUpdatedStudent = studentRepository.getByID(requestUpdate.id);
+        if(ToBeUpdatedStudent==null){
+            throw new StudentNotFound("Student not found");
+        }
         ToBeUpdatedStudent.setName(requestUpdate.name);
         ToBeUpdatedStudent.setAge(requestUpdate.age);
         ToBeUpdatedStudent.setScore(requestUpdate.score);
